+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Data files can be in Markdown (as well as JSON or YAML)
+
 ## [0.2.0] - 2021-07-22
 
 ### Changed